+++ packages/mongodb-schema/src/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './jobInterface.js'
 export * from './userInterface.js'
 export * from './multijobInterface.js'
+export * from './openmmInterface.js'